change some icons
