+++ src/components/Shared/MobileNav.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useState, Fragment } from "react";
+import React, { useContext, useState } from "react";
 import { TOOLS_PROVIDER } from "../../context/ToolsProvider";
 import avatar from "../../assets/image/avatar.svg";
 import { IoClose } from "react-icons/io5";
@@ -6,27 +6,20 @@ import { HiChevronRight, HiLogout } from "react-icons/hi";
 import { AUTH_CONTEXT } from "../../context/AuthProvider";
 import { Link } from "react-router-dom";
 import useAdmin from "../../hooks/useAdmin";
-import { useCart } from "react-use-cart";
-// logout modal
-import { Dialog, Transition } from "@headlessui/react";
+
+
+
+import LogoutModal from "../modal/LogoutModal";
 
 const MobileNav = () => {
   const [openModal, setOpenModal] = useState(false);
-
   const { openMenu, setOpenMenu } = useContext(TOOLS_PROVIDER);
-  const { user, logOut } = useContext(AUTH_CONTEXT);
-  const { emptyCart } = useCart();
+  const { user } = useContext(AUTH_CONTEXT);
   const [isAdmin] = useAdmin(user?.email);
 
   const handleClickLogout = () => {
     setOpenModal(true);
     setOpenMenu(false);
-  };
-  const handleLogOut = () => {
-    emptyCart();
-    localStorage.removeItem("accessToken");
-    logOut();
-    setOpenModal(false);
   };
 
   return (
@@ -146,65 +139,7 @@ const MobileNav = () => {
       </div>
       {/* Logout modal open */}
       {openModal && (
-        <Transition appear show={openModal} as={Fragment}>
-          <Dialog
-            as="div"
-            className="relative z-50"
-            onClose={() => setOpenModal(false)}
-          >
-            <Transition.Child
-              as={Fragment}
-              enter="ease-out duration-300"
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leave="ease-in duration-200"
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <div className="fixed inset-0 bg-black bg-opacity-25" />
-            </Transition.Child>
-
-            <div className="fixed inset-0 overflow-y-auto">
-              <div className="flex min-h-full items-center justify-center p-4 text-center">
-                <Transition.Child
-                  as={Fragment}
-                  enter="ease-out duration-300"
-                  enterFrom="opacity-0 scale-95"
-                  enterTo="opacity-100 scale-100"
-                  leave="ease-in duration-200"
-                  leaveFrom="opacity-100 scale-100"
-                  leaveTo="opacity-0 scale-95"
-                >
-                  <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
-                    <Dialog.Title
-                      as="h3"
-                      className="text-lg font-medium leading-6 text-gray-900"
-                    >
-                      Are you sure you want to log out of your account?
-                    </Dialog.Title>
-
-                    <div className="mt-4 flex items-center justify-end space-x-3">
-                      <button
-                        type="button"
-                        className="inline-flex justify-center rounded-md border border-transparent bg-neutral-800 px-4 py-2 text-sm font-medium text-white hover:bg-neutral-900  "
-                        onClick={handleLogOut}
-                      >
-                        Yes
-                      </button>
-                      <button
-                        type="button"
-                        className="inline-flex justify-center rounded-md border border-transparent bg-neutral-200 px-4 py-2 text-sm font-medium text-neutral-800 hover:bg-neutral-300 "
-                        onClick={() => setOpenModal(false)}
-                      >
-                        No
-                      </button>
-                    </div>
-                  </Dialog.Panel>
-                </Transition.Child>
-              </div>
-            </div>
-          </Dialog>
-        </Transition>
+        <LogoutModal openModal={openModal} setOpenModal={setOpenModal} />
       )}
     </>
   );
