+++ src/components/Shared/icons/svgIcons.jsx
@@ -1,0 +1,532 @@
+export const userIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={24}
+      height={24}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M154.666 117.333H170.667C194.230 117.333 213.333 136.436 213.333 160.000V192.000C213.333 203.782 203.782 213.333 192.000 213.333H64.000C52.218 213.333 42.667 203.782 42.667 192.000V160.000C42.667 136.436 61.769 117.333 85.333 117.333H101.334"
+          stroke="#000000"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="primary"
+        />
+        <path
+          d="M170.667 82.526C170.667 110.431 146.963 128.000 128.000 128.000C109.037 128.000 85.333 110.431 85.333 82.526C85.333 54.621 101.818 32.000 128.000 32.000C154.182 32.000 170.667 54.621 170.667 82.526Z"
+          stroke="url(#a)"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="secondary"
+        />
+      </g>
+    </svg>
+  );
+};
+export const orderIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={24}
+      height={24}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M213.333 160.000V53.333C213.333 41.551 203.782 32.000 192.000 32.000H64.000C52.218 32.000 42.667 41.551 42.667 53.333V202.667C42.667 214.449 52.218 224.000 64.000 224.000H149.333M213.333 160.000L149.333 224.000M213.333 160.000H170.667C158.884 160.000 149.333 169.551 149.333 181.333V224.000"
+          stroke="#000000"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="primary"
+        />
+        <path
+          d="M85.333 74.667H170.667M85.333 117.333H149.333M85.333 160.000H117.333"
+          stroke="url(#a)"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="secondary"
+        />
+      </g>
+    </svg>
+  );
+};
+export const settingIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={24}
+      height={24}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M106.667 32.000H149.333V56.425C154.299 57.903 159.049 59.884 163.522 62.308L180.797 45.033L210.967 75.203L193.692 92.478C196.116 96.952 198.097 101.701 199.574 106.667H224.000V149.333H199.574C198.097 154.299 196.116 159.049 193.692 163.522L210.967 180.797L180.798 210.967L163.522 193.692C159.049 196.116 154.299 198.097 149.333 199.574V224.000H106.667V199.574C101.701 198.097 96.951 196.116 92.478 193.692L75.202 210.967L45.033 180.798L62.308 163.522C59.884 159.049 57.903 154.299 56.425 149.333H32.000V106.667H56.425C57.903 101.701 59.884 96.951 62.308 92.478L45.033 75.203L75.203 45.033L92.478 62.308C96.952 59.884 101.701 57.903 106.667 56.425V32.000Z"
+          stroke="#000000"
+          strokeWidth={14}
+          strokeLinejoin="round"
+          id="primary"
+        />
+        <path
+          d="M160.000 128.000C160.000 145.674 145.674 160.000 128.000 160.000C110.326 160.000 96.000 145.674 96.000 128.000C96.000 110.326 110.326 96.000 128.000 96.000C145.674 96.000 160.000 110.326 160.000 128.000Z"
+          stroke="url(#a)"
+          strokeWidth={14}
+          strokeLinejoin="round"
+          id="secondary"
+        />
+      </g>
+    </svg>
+  );
+};
+export const dashboardIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={24}
+      height={24}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M32.000 74.667H96.000M224.000 74.667H96.000M96.000 74.667V224.000"
+          stroke="url(#a)"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="secondary"
+        />
+        <path
+          d="M202.667 32.000H53.333C41.551 32.000 32.000 41.551 32.000 53.333V74.667V202.667C32.000 214.449 41.551 224.000 53.333 224.000H96.000H202.667C214.449 224.000 224.000 214.449 224.000 202.667V74.667V53.333C224.000 41.551 214.449 32.000 202.667 32.000Z"
+          stroke="#000000"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="primary"
+        />
+      </g>
+    </svg>
+  );
+};
+export const logoutIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={24}
+      height={24}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M128.000 42.667H85.333C73.551 42.667 64.000 52.218 64.000 64.000L64.000 192.000C64.000 203.782 73.551 213.333 85.333 213.333H128.000"
+          stroke="#000000"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="primary"
+        />
+        <path
+          d="M202.667 128.000H106.667M202.667 128.000L170.667 96.000M202.667 128.000L170.667 160.000"
+          stroke="url(#a)"
+          strokeWidth={14}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="secondary"
+        />
+      </g>
+    </svg>
+  );
+};
+export const cartIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={32}
+      height={32}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={25}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={25}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={25}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={25}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <rect id="backgr" width={256} height={256} fill="none" rx={60} />
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M170.667 106.667V85.333C170.667 61.769 151.564 42.667 128.000 42.667C104.436 42.667 85.333 61.769 85.333 85.333V106.667"
+          stroke="#ff6600"
+          strokeWidth={25}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="secondary"
+        />
+        <path
+          d="M53.333 128.000C53.333 116.218 62.885 106.667 74.667 106.667H181.333C193.116 106.667 202.667 116.218 202.667 128.000V170.667C202.667 194.230 183.564 213.333 160.000 213.333H96.000C72.436 213.333 53.333 194.230 53.333 170.667V128.000Z"
+          stroke="#000000"
+          strokeWidth={25}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="primary"
+        />
+      </g>
+    </svg>
+  );
+};
+export const cartIconWhite = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={32}
+      height={32}
+      viewBox="0 0 256 256"
+      fill="none"
+      id="my-svg"
+    >
+      <defs>
+        <pattern
+          id="a"
+          patternUnits="userSpaceOnUse"
+          width={80}
+          height={80}
+          patternTransform="scale(3.19) rotate(0)"
+        >
+          <rect
+            x={0}
+            y={0}
+            width="100%"
+            height="100%"
+            fill="hsla(0,0%,100%,1)"
+          />
+          <path
+            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ffc400"
+            fill="none"
+          />
+          <path
+            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff8c00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#ff2a00"
+            fill="none"
+          />
+          <path
+            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
+            strokeWidth={30}
+            stroke="#cf0000"
+            fill="none"
+          />
+        </pattern>
+        <linearGradient id="gradient1">
+          <stop className="stop1" offset="0%" stopColor="#8f66ff" />
+          <stop className="stop2" offset="100%" stopColor="#3d12ff" />
+        </linearGradient>
+      </defs>
+      <rect id="backgr" width={256} height={256} fill="none" rx={60} />
+      <g id="group" transform="translate(0,0) scale(1)">
+        <path
+          d="M170.667 106.667V85.333C170.667 61.769 151.564 42.667 128.000 42.667C104.436 42.667 85.333 61.769 85.333 85.333V106.667"
+          stroke="#ff8000"
+          strokeWidth={25}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="secondary"
+        />
+        <path
+          d="M53.333 128.000C53.333 116.218 62.885 106.667 74.667 106.667H181.333C193.116 106.667 202.667 116.218 202.667 128.000V170.667C202.667 194.230 183.564 213.333 160.000 213.333H96.000C72.436 213.333 53.333 194.230 53.333 170.667V128.000Z"
+          stroke="#ffffff"
+          strokeWidth={25}
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          id="primary"
+        />
+      </g>
+    </svg>
+  );
+};
