+++ src/components/products/Product.jsx
@@ -1,20 +1,16 @@
 import React, { useContext, useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { HiOutlinePhoto } from "react-icons/hi2";
-import {
-  HiCurrencyBangladeshi,
-  HiOutlineShoppingBag,
-  HiCheckCircle,
-} from "react-icons/hi";
+import { HiCurrencyBangladeshi, HiCheckCircle } from "react-icons/hi";
 import { useCart } from "react-use-cart";
 import useFavorite from "../../hooks/useFavorite";
 import { AUTH_CONTEXT } from "../../context/AuthProvider";
 import { useQuery } from "@tanstack/react-query";
-
+import cartIcon from "../../assets/icons/cart.svg";
+import addCartIcon from "../../assets/icons/cart-add.svg";
 const Product = ({ product }) => {
   const [imageLoad, setImageLoad] = useState(false);
   const [isVisible, setIsVisible] = useState(false);
-  const [isFav, setIsFav] = useState(false);
   const { user } = useContext(AUTH_CONTEXT);
   const { name, img, price, _id, outOfStock } = product;
   const { addItem } = useCart();
@@ -193,286 +189,17 @@ const Product = ({ product }) => {
               {/* mobile btn */}
               <button
                 onClick={handleAddToCart}
-                className="bg-neutral-s100 inline-flex items-center hover:bg-neutral-200 rounded-full transition duration-300 p-1 text-xs"
+                className="inline-flex items-center rounded-full p-1 text-xs"
                 title="Add to cart"
               >
                 <span className="text-black p-1">
                   {isVisible ? (
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width={32}
-                      height={32}
-                      viewBox="0 0 256 256"
-                      fill="none"
-                      id="my-svg"
-                    >
-                      <defs>
-                        <pattern
-                          id="a"
-                          patternUnits="userSpaceOnUse"
-                          width={80}
-                          height={80}
-                          patternTransform="scale(3.19) rotate(0)"
-                        >
-                          <rect
-                            x={0}
-                            y={0}
-                            width="100%"
-                            height="100%"
-                            fill="hsla(0,0%,100%,1)"
-                          />
-                          <path
-                            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#ffc400"
-                            fill="none"
-                          />
-                          <path
-                            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#ff8c00"
-                            fill="none"
-                          />
-                          <path
-                            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#ff2a00"
-                            fill="none"
-                          />
-                          <path
-                            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#cf0000"
-                            fill="none"
-                          />
-                        </pattern>
-                        <linearGradient id="gradient1">
-                          <stop
-                            className="stop1"
-                            offset="0%"
-                            stopColor="#8f66ff"
-                          />
-                          <stop
-                            className="stop2"
-                            offset="100%"
-                            stopColor="#3d12ff"
-                          />
-                        </linearGradient>
-                      </defs>
-                      <rect
-                        id="backgr"
-                        width={256}
-                        height={256}
-                        fill="none"
-                        rx={60}
-                      />
-                      <g id="group" transform="translate(0,0) scale(1)">
-                        <path
-                          d="M64.000 53.333H224.000L213.333 160.000H85.333L64.000 53.333ZM64.000 53.333L53.333 32.000H32.000M117.333 208.000C117.333 216.836 110.170 224.000 101.333 224.000C92.497 224.000 85.333 216.836 85.333 208.000C85.333 199.164 92.497 192.000 101.333 192.000C110.170 192.000 117.333 199.164 117.333 208.000ZM213.333 208.000C213.333 216.836 206.170 224.000 197.333 224.000C188.497 224.000 181.333 216.836 181.333 208.000C181.333 199.164 188.497 192.000 197.333 192.000C206.170 192.000 213.333 199.164 213.333 208.000Z"
-                          stroke="#000000"
-                          strokeWidth={14}
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          id="primary"
-                        />
-                        <path
-                          d="M128.000 106.667L149.333 128.000L181.333 85.333"
-                          stroke="url(#a)"
-                          strokeWidth={14}
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          id="secondary"
-                        />
-                      </g>
-                    </svg>
+                    <img src={addCartIcon} alt="" />
                   ) : (
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width={32}
-                      height={32}
-                      viewBox="0 0 256 256"
-                      fill="none"
-                      id="my-svg"
-                    >
-                      <defs>
-                        <pattern
-                          id="a"
-                          patternUnits="userSpaceOnUse"
-                          width={80}
-                          height={80}
-                          patternTransform="scale(3.19) rotate(0)"
-                        >
-                          <rect
-                            x={0}
-                            y={0}
-                            width="100%"
-                            height="100%"
-                            fill="hsla(0,0%,100%,1)"
-                          />
-                          <path
-                            d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#ffc400"
-                            fill="none"
-                          />
-                          <path
-                            d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#ff8c00"
-                            fill="none"
-                          />
-                          <path
-                            d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#ff2a00"
-                            fill="none"
-                          />
-                          <path
-                            d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                            strokeWidth={30}
-                            stroke="#cf0000"
-                            fill="none"
-                          />
-                        </pattern>
-                        <linearGradient id="gradient1">
-                          <stop
-                            className="stop1"
-                            offset="0%"
-                            stopColor="#8f66ff"
-                          />
-                          <stop
-                            className="stop2"
-                            offset="100%"
-                            stopColor="#3d12ff"
-                          />
-                        </linearGradient>
-                      </defs>
-                      <rect
-                        id="backgr"
-                        width={256}
-                        height={256}
-                        fill="none"
-                        rx={60}
-                      />
-                      <g id="group" transform="translate(0,0) scale(1)">
-                        <path
-                          d="M64.000 53.333H224.000L213.333 160.000H85.333L64.000 53.333ZM64.000 53.333L53.333 32.000H32.000M117.333 208.000C117.333 216.836 110.170 224.000 101.333 224.000C92.497 224.000 85.333 216.836 85.333 208.000C85.333 199.164 92.497 192.000 101.333 192.000C110.170 192.000 117.333 199.164 117.333 208.000ZM213.333 208.000C213.333 216.836 206.170 224.000 197.333 224.000C188.497 224.000 181.333 216.836 181.333 208.000C181.333 199.164 188.497 192.000 197.333 192.000C206.170 192.000 213.333 199.164 213.333 208.000Z"
-                          stroke="#000000"
-                          strokeWidth={14}
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          id="primary"
-                        />
-                        <path
-                          d="M149.333 85.333V106.667M149.333 106.667V128.000M149.333 106.667H128.000M149.333 106.667H170.667"
-                          stroke="url(#a)"
-                          strokeWidth={14}
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          id="secondary"
-                        />
-                      </g>
-                    </svg>
+                    <img src={cartIcon} alt="" />
                   )}
                 </span>
               </button>
-              {/* desktop btn */}
-              {/* <button
-                onClick={handleAddToCart}
-                className="bg-neutral-200 hidden lg:inline-flex items-center hover:bg-neutral-200 rounded-full transition duration-300 px-4 py-2 text-xs"
-                title="Add to cart"
-              >
-                <span className="text-black mr-1">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width={32}
-                    height={32}
-                    viewBox="0 0 256 256"
-                    fill="none"
-                    id="my-svg"
-                  >
-                    <defs>
-                      <pattern
-                        id="a"
-                        patternUnits="userSpaceOnUse"
-                        width={80}
-                        height={80}
-                        patternTransform="scale(3.19) rotate(0)"
-                      >
-                        <rect
-                          x={0}
-                          y={0}
-                          width="100%"
-                          height="100%"
-                          fill="hsla(0,0%,100%,1)"
-                        />
-                        <path
-                          d="M-20.133 4.568C-13.178 4.932-6.452 7.376 0 10c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                          strokeWidth={30}
-                          stroke="#ffc400"
-                          fill="none"
-                        />
-                        <path
-                          d="M-20.133 24.568C-13.178 24.932-6.452 27.376 0 30c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                          strokeWidth={30}
-                          stroke="#ff8c00"
-                          fill="none"
-                        />
-                        <path
-                          d="M-20.133 44.568C-13.178 44.932-6.452 47.376 0 50c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                          strokeWidth={30}
-                          stroke="#ff2a00"
-                          fill="none"
-                        />
-                        <path
-                          d="M-20.133 64.568C-13.178 64.932-6.452 67.376 0 70c6.452 2.624 13.036 5.072 20 5 6.967-.072 13.56-2.341 20-5 6.44-2.659 13.033-4.928 20-5 6.964-.072 13.548 2.376 20 5s13.178 5.068 20.133 5.432"
-                          strokeWidth={30}
-                          stroke="#cf0000"
-                          fill="none"
-                        />
-                      </pattern>
-                      <linearGradient id="gradient1">
-                        <stop
-                          className="stop1"
-                          offset="0%"
-                          stopColor="#8f66ff"
-                        />
-                        <stop
-                          className="stop2"
-                          offset="100%"
-                          stopColor="#3d12ff"
-                        />
-                      </linearGradient>
-                    </defs>
-                    <rect
-                      id="backgr"
-                      width={256}
-                      height={256}
-                      fill="none"
-                      rx={60}
-                    />
-                    <g id="group" transform="translate(0,0) scale(1)">
-                      <path
-                        d="M64.000 53.333H224.000L213.333 160.000H85.333L64.000 53.333ZM64.000 53.333L53.333 32.000H32.000M117.333 208.000C117.333 216.836 110.170 224.000 101.333 224.000C92.497 224.000 85.333 216.836 85.333 208.000C85.333 199.164 92.497 192.000 101.333 192.000C110.170 192.000 117.333 199.164 117.333 208.000ZM213.333 208.000C213.333 216.836 206.170 224.000 197.333 224.000C188.497 224.000 181.333 216.836 181.333 208.000C181.333 199.164 188.497 192.000 197.333 192.000C206.170 192.000 213.333 199.164 213.333 208.000Z"
-                        stroke="#000000"
-                        strokeWidth={14}
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        id="primary"
-                      />
-                      <path
-                        d="M149.333 85.333V106.667M149.333 106.667V128.000M149.333 106.667H128.000M149.333 106.667H170.667"
-                        stroke="url(#a)"
-                        strokeWidth={14}
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        id="secondary"
-                      />
-                    </g>
-                  </svg>
-                </span>
-                Add to cart
-              </button> */}
             </div>
           </div>
         </div>
